+++ src/WhereGroup/MetadorBundle/Controller/ServiceController.php
@@ -8,6 +8,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
+use WhereGroup\SearchBundle\Component\Paging;
 
 /**
  * @Route("/metador/service")
@@ -19,20 +20,17 @@ class ServiceController extends Controller
      */
     public function indexAction()
     {
-        $limit = 100;
-        $offset = 0;
-
+        $page     = $this->get('request')->get('page', 1);
         $metadata = $this->get('metador_metadata');
-
-        // Load Template.
-        $conf = $this->container->getParameter('metador');
-
-        $tmp = $metadata->getService($limit, $offset);
+        $conf     = $this->container->getParameter('metador');
+        $limit    = $conf['hits'];
+        $paging   = new Paging($metadata->getServiceCount(), $limit, $page);
 
         return $this->render(
             $conf['templates']['form'] . ':Service:index.html.twig',
             array(
-                'rows' => $metadata->getService($limit, $offset)
+                'rows'   => $metadata->getService($limit, ($page * $limit) - $limit),
+                'paging' => $paging->calculate()
             )
         );
     }
@@ -60,8 +58,8 @@ class ServiceController extends Controller
         return $this->render(
             $conf['templates']['form'] . ':Service:form.html.twig',
             array(
-                'p' => $p,
-                'examples' => $this->getExample(),
+                'p'         => $p,
+                'examples'  => $this->getExample(),
                 'hasAccess' => true
             )
         );
@@ -88,8 +86,8 @@ class ServiceController extends Controller
         return $this->render(
             $conf['templates']['form'] . ':Service:form.html.twig',
             array(
-                'p' => $p,
-                'examples' => $this->getExample(),
+                'p'         => $p,
+                'examples'  => $this->getExample(),
                 'hasAccess' => true
             )
         );
@@ -121,9 +119,9 @@ class ServiceController extends Controller
         return $this->render(
             $conf['templates']['form'] . ':Service:form.html.twig',
             array(
-                'id' => $id,
-                'p' => $p,
-                'examples' => $this->getExample(),
+                'id'        => $id,
+                'p'         => $p,
+                'examples'  => $this->getExample(),
                 'hasAccess' => !$data->getReadonly()
             )
         );