+++ app/Resources/doc/changelog.md
@@ -6,6 +6,7 @@
 * WMS GetCapabilities upload.
 * Added UUID component.
 * Search element for mapbender.
+* Added Paging.
 
 ## 2.0.5.0
 * XML-Import command.